backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
openai==1.3.5
pydantic==2.5.0
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
murf==1.0.0

